Like totally got rid of warnings
Which is actually necessary on my machine because
it would throw "invalid operation" errors whenever
a function lied about not returning anything.

I did some slashy stuff where I just turned the last
`else if` into an `else`, so if that becomes a problem
later we just need to find something to return at the
end in all cases. Although the behaviour should be
and appears the same.

tl;dr long ass commit, but everything works. For now.